Acomode la columna balance, de la tabla de users, ahora es double(5,2) y no acepta mas de 999.99 de balance,cree la funcion select en la libreria de mysql y le anadi la validacion si ya se creo o no la conexion con la base de datos mysql, agregue dicha validacion a la funcion insert, en la misma libreria. Cree la funcion getUserByIdMySQL en el servicio users.js y finalmente, en la ruta de auth.js, en el endpoint de /sign-in, agregue la logica para obtener el balance y el nivel del user, aparte de almacenarla en la response final del endpoint.

+++ utils/mysqlschema/hf_club_database01.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS users(
     user_id VARCHAR(24) PRIMARY KEY,
-    balance DOUBLE(4,2) DEFAULT 0.00,
+    balance DOUBLE(5,2) DEFAULT 0.00,
     active TINYINT(1) DEFAULT 0,
     level TINYINT(1) DEFAULT 1,
     banned TINYINT(1) DEFAULT 0,
